[이용호] 미션 성공 게임과 연동
미션이 성공했을 때 미션페이즈의 시간이 깎이도록 페이즈와 연동

+++ Source/Vigilant/Core/GamePhases/VGMissionPhase.h
@@ -24,7 +24,7 @@ public:
 	virtual void ResumePhase() override;
 	
 	virtual bool CanPlayerInteract(AVGCharacterBase* Player, AActor* InteractableObject) override;
-	virtual void OnMissionCleared(int32 TimeReducedAmount) override;
+	virtual void OnMissionCleared(float TimeReducedAmount) override;
 	
 	void OnMissionTimeUp();
 };
